third_party/blink/web_tests/platform/generic/fast/canvas/OffscreenCanvas-2d-imageData-in-worker-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

PASS createImageData(sx, sy) creates blank image correctly
PASS createImageData(ImageData) creates blank image correctly
PASS getImageData(sx, sy, sw, sh) is successful.
PASS putImageData(ImageData, dx, dy) is successful.
PASS putImageData(ImageData, dx, dy, dirtyX, dirtyY, dirtyWidth, dirtyHeight) is successful.
